fix: ipXtables: don't use tables that aren't available
At least for the default ruleset we can avoid failure if some of these
tables are missing. But features that use those missing tables will
still fail if the user attempts to use those features.

Here is a probably incomplete mapping of tables -> features:

  raw:	    helpers, IPv6_rpfilter
  mangle:   rich rule mark action
  nat:	    masquerade, forward ports
  security: none

Of course, direct rules apply to all tables. It is fatal if the "filter"
table is not available.

Fixes: #411
Fixes: #484