fix: ipXtables: using "mangle" in zone not dependent on "nat"
Since we removed the use of "mangle" in forward ports, the use of
"mangle" in a zone is no longer dependent on "nat" also being usable.
The only thing that uses "mangle" at the moment is the rich rule mark
action, which is not dependent on other tables.

Fixes: 7235f76a8183 ("improvement: iptables: simplify forward port rules")
(cherry picked from commit b08695e622cc5d586d675384be7c80c0222a39bf)