improvement: nftables: only add helper object if adding the port/rule
No noticeable change. We were also adding them on a delete, but nftables
handles the duplicate add without error. We're still not deleting the
object on a service delete because that would require reference counting
as multiple services use the same helper. It's not worth the complexity
when we can simply leave the object around. Users can --reload if they
really want to get rid of them.