chore: travis: no longer run testsuite
It has become increasingly difficult to keep the testsuite functional in
travis-ci. Many of the tests are are skipped due to missing kernel
features or bugs. The nftables backend requires kernel >4.18. At the
moment the latest travis-ci offers is Ubuntu Bionic (4.15).

From now on, we're only doing source code and build checks.