fix: do not allow zone drifting
Chain zone dispatch together and always use "goto". This guarantees
there is no fall through to other zones. This was especially problematic
in regards to the default zone.

This removes the _ZONES_SOURCE chains, but adds _ZONES_IFACES. At the
end of _ZONES we do a goto to _ZONES_IFACES. This is so sources always
take precedence over interfaces.

Fixes: #258
Fixes: #441