fix: dbus: fix service API break
This fixes a dbus API break that occurred when introducing service
includes. The includes were added to the method's tuple, but doing so
changed the dbus signature and thus broke the API. This restores the old
signature.

Move to using key,value based import/export and sanity checking.
Previously we were using a tuple with semi-undocumented positions.

Fixes: 1fc208bf9317 ("feat: service includes")
Fixes: rhbz 1721414