fix: guarantee zone source dispatch is sorted by zone name
Apparently users depend on firewalld sorting zone dispatch for sources
by the zone name. This is used to specify precedence for overlapping
address spaces.

Since we have to track rule positions of source based dispatch we might
as well abuse this and combine the source/interface dispatch into a
single chain.

Fixes: rhbz 1734765
Fixes: 70993581d79b ("fix: do not allow zone drifting")