fix(translation): several translations have a bad value for "Interfaces"
These translations are obviously incorrect. Resetting them to the empty
string indicates to translators that new translations need to be
provided.

Per
<https://www.gnu.org/software/gettext/manual/html_node/Untranslated-Entries.html>,
"When xgettext originally creates a PO file, [...] it initializes the
msgid field with the untranslated string, and leaves the msgstr string
to be empty. Such entries, having an empty translation, are said to be
untranslated entries."

Alleviates the symptoms of #499 but not the cause.