fix: Revert "ebtables: drop support for broute table"
This reverts commit 0e78dea21ffb03ea2e51df6158471b9cbd6652c9.

Apparently there _are_ users of this table. Let's revert this change and
be a little smarter about not attempting to use the table if it's not
available (ebtables-nft). We'll still fail if the user has direct rules
for this table, but at least the default ruleset will work.

(cherry picked from commit ca34c9e051807d62371d5e980210f11859cab360)