fix/cleanup: relax auth for property getters to PK_ACTION_CONFIG_INFO
On FirewallD, the getters were missing an explicit annotation.  The default
is PK_ACTION_CONFIG.  All other methods on FirewallD are annotated with
an explicit auth requirement.  So lets be explicit about the auth
requirement for the FirewallD getters.

And lets change the requirement (for the getters) to PK_ACTION_CONFIG_INFO.
This was the old default, before CVE-2016-5410[1].

[1] https://www.openwall.com/lists/oss-security/2016/08/16/3

It should be _CONFIG_INFO not just _INFO, because it discloses
"nf_conntrack_helper_setting".  Also, the setter is marked as _CONFIG.

Apply the same change to the config object and all the config sub-objects.
They also have setters which require PK_ACTION_CONFIG.