improvement: Select target chain in rich rules
Adds needed syntax, implementation in ipXtables and nftables,
documentation, and tests to extend the rich rule syntax to allow
rules to be added to chains other than INPUT. Currently limited
to INPUT, FORWARD_IN, and FORWARD_OUT. OUTPUT could be added easily as
soon as firewalld creates the needed per-zone OUTPUT chains.

Currently, firewalld lacks the ability to be effectively used on a
router due to rules only getting added to the INPUT chain. This commit
seeks to resolve this.