allow forwarding between interfaces in a zone
This introduces a new boolean flag called "forward" on zones, similar to
the existing "masquerade" setting.

When the "forward" setting is enabled for a zone, the firewall rules
will allow packets to be forwarded between interfaces within that zone,
but not to other zones. Unlike the "masquerade" setting, no other NAT
rules are applied.

This gives a rudimentary kind of configurable forwarding. You can
control whether packets will be forwarded by arranging your interfaces
into zones and turning on this flag.