Use relative python shebangs
When cross-compiling, Python is built first and installed to a staging
directory and the target. When cross-compiling firewalld, the Python shebangs
end up with a full, hard-coded path to the specified python interpreter,
IE: #!/home/buildroot/output/host/bin/python.

The following files are affected by this:
config/lockdown-whitelist.xml.in
src/firewall-applet.in
src/firewall-cmd.in
src/firewall-config.in
src/firewall-offline-cmd.in
src/firewalld.in
src/tests/atlocal.in

This full path causes the above files not to run, as the full cross-compiled
path does not exist on the cross-compiled device.

Instead of using #!@PYTHON@ use #!/usr/bin/env python@PYTHON_VERSION@ instead.
This change allows for the python path to be relative but still work with the
AM_PATH_PYTHON macro.