fix(client): only pass runtime zone settings to setZoneSettings2
The new setZoneSettings2 D-Bus method only expects runtime settings, it
does not handle some zone attributes which can only be changed in the
permanent settings: version, short, description, and target.

This was causing the setZoneSettings client method to fail, when it
tried to feed back the complete zone settings:

    Traceback (most recent call last):
      File "src/firewall/server/decorators.py", line 67, in dbus_handle_exceptions
        return func(*args, **kwargs)
      File "src/firewall/server/firewalld.py", line 917, in setZoneSettings2
        self.fw.zone.set_config_with_settings_dict(zone, settings, sender)
      File "src/firewall/core/fw_zone.py", line 398, in set_config_with_settings_dict
        raise FirewallError(errors.INVALID_SETTING, "Unhandled setting type {} key {}".format(type(settings[key]), key))
    firewall.errors.FirewallError: INVALID_SETTING: Unhandled setting type <class 'dbus.String'> key description

This introduces a new pair of client zone methods,
getRuntimeSettingsDict() / getRuntimeSettingsDbusDict()
which are analogous to the existing
getSettingsDict() / getSettingsDbusDict() methods, but exclude
attributes that cannot be changed in the runtime configuration.

Fixes: 67d057508f81 ("improvement(dbus): new dict based APIs for zones")