fix(LastUpdatedOrderedDict): __getitem__(): fetch from list if int
If the LastUpdatedOrderedDict contains a boolean key, e.g.

    myLastUpdatedOrderedDict = LastUpdatedOrderedDict()
    myLastUpdatedOrderedDic[True] = "true"

then

    myLastUpdatedOrderedDic[1]

yields "true". As such, using the LastUpdatedOrderedDict as an iterable
e.g.

    for foo in myLastUpdatedOrderedDict:
        ...

would mean that the for loop tries integer indexes 0 (returns key True),
and then 1 (also returns key True). This caused duplicate walks of a key
True if it was the first key in the LastUpdatedOrderedDict.

This occurs because

    >>> True == 1
    True
    >>> False == 0
    True