fix(policy): zone interface/source changes should affect all using zone
This was caused by bad usage of lambda. The lambda's were added to the
post hooks of the transaction, but the closure caused all of them to use
the same policy value - that is, the value as of the last iteration of
the for loop.

To fix add an argument to the lambda and pass the _current_ value of
policy to the post hook of the transaction.