fix(zone): target: default is now similar to reject
The only difference is default implicitly allows ICMP.

BREAKING CHANGE:

Firstly, this affects a combination in which the ingress zone uses a
default target and the egress zone uses ACCEPT. In this scenario packets
would "fall through" and the egress zone would be responsible for the
decision to accept/reject. A good example is public --> trusted. This is
now disallowed. It's unexpected and a security concern.

If users desire the old behavior it can be re-added via a policy.

  # firewall-cmd --permanent --new-policy allowForward
  # firewall-cmd --permanent --policy allowForward --set-target ACCEPT
  # firewall-cmd --permanent --policy allowForward --add-ingress-zone public
  # firewall-cmd --permanent --policy allowForward --add-egress-zone trusted

Secondly this has implications for things that occur after the
filter/forward chain: masquerade, tcp-mss-clamp. In the past the default
target (e.g. public zone) would allow the traffic to "fall through" to
zones that may be using these features. This is very similar to the
issue mentioned directly above and has the same cause. It is now
disallowed.

Note: masquerade is unaffected for ingress zones that use ACCEPT. As
such if your LAN is in the internal or trusted zone and your uplink is
in the external zone then it will still masquerade as expected. This
change only affects the case when ingress zone target == default.

If users desire the old behavior it can be re-added via a policy.

  # firewall-cmd --permanent --new-policy masqueradePolicy
  # firewall-cmd --permanent --policy masqueradePolicy --add-ingress-zone public
  # firewall-cmd --permanent --policy masqueradePolicy --add-egress-zone external
  # firewall-cmd --permanent --policy masqueradePolicy --add-masquerade

or for tcp-mss-clamp,

  # firewall-cmd --permanent --new-policy clampTcpMss
  # firewall-cmd --permanent --policy clampTcpMss --add-ingress-zone public
  # firewall-cmd --permanent --policy clampTcpMss --add-egress-zone external
  # firewall-cmd --permanent --policy clampTcpMss --add-rich-rule='rule tcp-mss-clamp'

Fixes: #177