test(dbus): auth: verify we check UID for dbus calls
Running without polkit is a valid scenario. When it's not available we
use dbus to check if the sender's UID == 0. Lets make sure the daemon is
actually checking UID == 0.

Note: This is an integration test because some distributions have some
form of sudo and/or dbus interaction with systemd-logind. logind is not
available via dbus in our isolated/namespace testsuite. If logind is not
reachable we'll get dbus timeouts.