test(functions): set defaults for values in atlocal
Covers the case that the testsuite is run from a location in which
atconfig and atlocal can't be found. This is the case if the testsuite
has been installed or run from a location other than the source tree.

(cherry picked from commit e1041a0a349c914e9ebdea03a5ab387b6c5e8ea1)