fix(policy): unreachable when priority > 0
Since f2896e43c3a5 ("fix(zone): target: default is now similar to
reject") zones are fully terminal. They always end in an
accept/drop/reject. As a result policies with a priority >0 are not
actually reachable. The dispatch happens _after_ the zone's chains.

The solution is to jump to the policy dispatch just before the catch-all
accept/drop/reject in the zone. Since we're doing that we might as well
move the < 0 policy dispatch into the zone chains for uniformity. The
latter has no functional difference.

The exception is the OUTPUT chains. They don't have zone dispatch so the
top level jump to policy dispatch is retained.

Fixes: f2896e43c3a5 ("fix(zone): target: default is now similar to reject")