chore(rpm): firewall-config: add Recommends polkit
firewalld (the daemon) does polkit authorization. This is mostly
intended for workstation use cases, e.g. a desktop environment is
installed, where actions may be done by a non-root user. As such add the
Recommends to firewall-config.

polkit is optional. When it's not available firewalld falls back to UID
based authorization.

Fixes: rhbz 1931966