fix(firewalld): keep linux capability CAP_SYS_MODULE
When firewalld calls ip6tables it may implicitly load the ip6_tables, et
al kernel modules. As such we need to retain CAP_SYS_MODULE so that
implicit module is allowed. Otherwise we get EPERM from the kernel.

This only affects the -legacy variants and the top level table/chain
modules. The userspace binaries will modprobe the kernel modules.
Extensions, e.g. xt_conntrack, are implicitly loaded by the kernel based
on the rules being added and thus not subject to linux capabilities
checks.

The -nft variants are unaffected because they use the nftables
infrastructure which has implicit module loading in the kernel similar
to the iptables extensions (xt_* modules).

Fixes: rhbz 1990271
Fixes: fb0532e8a200 ("feat(firewalld): drop linux capabilities")