Revert "fix(service): UPnP Client: actually allow SSDP traffic"
This reverts commit 28fe70a4d70332edfe95cc6404f455a9b11ee89b.

--

SSDP uses multicast and a known UDP port. There is no in kernel
connection tracking support for SSDP. However, there is conntrackd
support. Unfortunately firewalld lacks conntrackd support at the moment.

For the above reasons we need to hold on to the source-port match in
this service definition. It's not the most secure, but it's the only way
to allow ssdp replies.

Reported-by: libera.chat user feral_hedgehog