fix(io/functions): check_config against on disk conf
Before this change the runtime FirewallConfig() instance was used. This
caused some permanent configuration issues to not be caught due to
comparing against the runtime instances of all objects.

For example, two zones in permanent configuration may use the same
interface (which is not valid), but if the runtime configuration does
not have have these interface assignments then check_config() won't
catch the issue since it compares against the runtime configuration.

Fix is to build a temporary FirewallConfig() instance for all the
on-disk/permanent configuration.