fix(fw_icmptype): don't alter ipvs in io_object
We should never modify the runtime icmptype object. This caused sanity
failures on --runtime-to-permanent because the runtime icmptype
pretended to be ipv4 only, but our configuration (policy
allow-host-ipv6) used it inside an IPv6 only rich rule.

For users that disable IPv6 then this probably triggered an accidental
copy of the stock icmptype to a user icmptype, but with the ipv6 family
dropped.