fix(runtimeToPermanent): errors for interfaces not in zone
We should only consider the interfaces for the currently iterated over
zone. Otherwise we will attempt to remove an interface from a zone for
which it is does not belong.

Note this only occurs when NetworkManager is running.

Fixes: #976
Fixes: rhbz2112982
(cherry picked from commit 15f47354c4a078dc694df1541550b3e5156548fc)