improvement(service): Zabbix Server: Rename service, update description
The service port that is opened by this is actually used by two Zabbix
components, the server and proxy, and the handler process is referred to
as the Zabbix Trapper, so rename the service and short description of
it.

Also update the long form description to describe what the port is used
for by Zabbix rather than what Zabbix is.