refactor(service): Rewrite Foreman, Foreman Proxy & Satellite
Foreman itself only runs on HTTP and HTTPS. It is the Foreman Proxy
which can talk to additional services such as DNS, DHCP, TFTP and
Puppet. However, all those services are optional.

This changes the services to describe just foreman and foreman-proxy. A
new foreman-proxy-optional service is created to list all optional
services that can be integrated. The optional aspect doesn't include the
foreman-proxy service itself since the port can differ between vanilla
Foreman (8443) and Foreman with Katello (9090).

At the same time it rewrites new RH-Satellite-6 and
RH-Satellite-6-capsule.

For RH-Satellite-6 port 5000 hasn't been in use since the switch to Pulp
3 in Satellite 6.10 (when crane was integrated in Pulp itself). Port
8080 hasn't been in use since candlepin listens on localhost on port
8443 instead. The Smart Proxy doesn't listen on port 8000 anymore, since
the Foreman Proxy templates feature was disabled by default. The other
listed ports (5646, 5647, 5671) are for qpid which was needed for
katello-agent. That has become optional and dropped with Satellite 6.14.

The RH-Satellite-6-capsule no longer includes RH-Satellite-6 because the
relation isn't there. Instead it lists the services it runs explicitly.