feat(zone): add support for priority
This adds support for specifying an ingress and egress priority for
zones. This allows customizing the order in which traffic is classified
into zones.

A zone with a lower priority value (higher precedence) will always be
considered before a zone with a higher priority value (lower
precedence). This is very similar to priority as used in policies and
rich rules.

Classification is essentially a sorted pair of (ingress priority, egress
priority). When priority is the same classification is sorted by
existing behavior, e.g. source based before interface based.

Fixes: #192
Fixes: rhbz2130288