chore(testsuite): rename "firewalld_test.py" to "firewalld_misc.py"
These are not unit tests, as they require a running (non-mock) firewalld
daemon.  They also interact with the host system by talking to the
"real" firewalld daemon. `make check` alleviates that by running those
processes in a separate namespace and D-Bus session. That's what the
caller can do, but the script itself is not a mostly self contained
"unit test", but rather integration tests.

On the other hand, consider that -- even if our tests use the unittest
module and not `pytest` -- it would be convenient to simply run the unit
tests with a plain `pytest` call. This would  automatically find the
tests in the file tree based on the file's names.

By default, pytest will run files named `*_test.py`. It would thus run
`firewalld_test.py`. That is wrong, as this is not a unit test that can
run without further, external setup. Rename `firewalld_test.py` so
that the file is not treated by pytest as a unit test.

Alternatively, we could set `[pytest].python_files` pattern in
`pytest.ini` or `pyproject.toml` to control which file names are
considered to be tests.  It seems however better to just follow the
default naming style and rename.