build(fedora): add "configure-for-system.sh" script
This script configures firewalld the same way, as during our RPM package
build. The purpose is, if you configure it this way, and then type `make
install`, that the firewalld files from your system are overwritten.

This is not something you would usually want. Usually you want that the
files built from source don't mess with the installed packages. This is
however useful in a VM or inside a container, where we want this.

The benefit is, after `make install`, we can just `systemctl
daemon-reload && systemctl restart firewalld` and have the new version
running.