improvement(functions): add ipaddr_{parse,norm}() function for normalizing
Add helper functions for parsing, verifying and normalizing
IP addresses (in string form).

We already have part of that functionality as checkIP(), checkIP6(), normalizeIP6().
Extend the scheme with the basic operations:

 - parse() (from string to a detailed representation)
 - unparse() (from the detailed representation to a normalized string)
 - norm() (is unparse(parse()))
 - check() (like parse(), but return a boolean instead of a ValueError)

There is some overlap with checkIP() and checkIP6(). Note that I am
mainly interested in the parsing/normalization here, hence the addition
of the new API.

The new API looks (from a naming point of view) different from
checkIP(). The reason is that Python commonly uses snake_case for method
names, and so does large parts of firewalld. Also, the name checkIP()
doesn't seem great for adding a parseIP() and normIP(). While it follows
the English verb-object structure, it seems clearer when the object
comes before the verb, so that it groups similar API by what they
operate on (ipaddr) instead of what they do (parse).