improvement(functions): use ipaddr_check() for checkIP()
There is little change further behavior by using the ipaddr_check()
functions. It internally still uses socket.inet_pton() for validating the
string. The point is that the "ipaddr_*()" functions are the basic
implementation for how to handle IP addresses as strings, and that
checkIP() should not implement it differently.