improvement(functions): use ipaddr_norm() for normalizeIP6()
- stripping "[]" from the address also accepts bogus input like
  "[[::" or "[]::1". Let's be strict here. Of course, this is a
  change in behavior, if you configured invalid IP addresses.

- normalizeIP6() should bring the IPv6 address in a canonical
  form. For example, "::0" and "0:00::" should all be normalized
  to "::".

- normalizeIP6() should raise an exception for bogus input. Previously,
  it would not, for example normalizeIP6("foo[") just returned "foo".

Use ipaddr_norm() for that. Since the parsing is now stricter, this can
break existing configurations. This is a change in behavior! But it
shouldn't break any correct configurations.

As normalizeIP6() now verifies the IP address (and raises a
ValueError()) otherwise, we also need to adjust checkIP6() to not emit
that exception. Do that by just calling ipaddr_check(), which is the
point of that function.