improvement(errors): add BugError exception class
Add an exception type that we can raise when we encounter a bug.
Maybe we could use AssertionError, but plain `assert` can be "optimized"
out in optimized mode. So while the habit of `raise AssertionError(...)`
may be fine, `assert [...]` is less fine.

To make the distinction clearer, add a special BugError exception type.