improvement(ipset): add ipset_entry_parse() function for parsing/validation
This aims to be like firewall.core.io.ipset.IPSet.check_entry(), but
instead of just checking, it also parses the entry and returns
information that can be used to process the parsed entry further.

This will later be called by firewall.core.io.ipset.IPSet.check_entry().

ipset_entry_parse() aims to parse everything that check_entry()
accepted, but in practice, they are not exactly the same. So a user
could have created entries with older versions of firewalld, which will
be considered invalid now. But those entries really should be invalid,
and not part of usable configuration.