improvement(build): add ".flake8" file for configuring "flake8"
flake8 is currently hooked up via `make -C src check-local`,
which is also called by ".github/workflows/source-checks.yml".

We suppress certain warnings and errors for our code base. That list of
suppressions is currently is only accessible via `make -C src
check-local`.

Add a `.flake8` file, so that running plain flake8 in the source tree
does the right thing.

Note that thereby we pass patterns like ".py", so if you add a new
python script to the source tree (which is neither in git nor in the
DIST list of make), it will be checked too. You avoid that, by keeping
the source directory clean. Otherwise, it seems desirable to check all
the files that are present.