improvement(ipset): use ipset_entry_parse() in IPSet.check_entry()
Reuse firewall.core.ipset.ipset_entry_parse() to validate the
IPSet entry in check_entry(). The point is to not only have a validation
function, but also a parse function so we can understand the entries.
Then the parse function is reused for the validation too.

With this, it would also be trivial to normalize the strings, which may
be a good idea.