improvement(client): add keyword arguments to FirewallClient._signal_receiver()
_signal_receiver() has a few direct callers, and is also passed as a
callback to bus.add_signal_receiver(). As such, the keyword arguments
are well known (they are "member", "interface" and "path"). Name them
explicitly instead of a **kwargs.

Also, no need to check whether "member" and "interface" arguments are
provided. All callers do, and if they would not, we can fail with an
exception.