improvement(client): remove unnecessary args argument from FirewallClient.connect()
None of the caller actually passes additional arguments.

Also, this is python, if you want to pass additional argument then do

  lambda *a: my_fcn(extra_arg, *a)

Most importantly, the callback is invoked by _signal_receiver, which
receives a *dbus_args list. For example, for the InterfaceAdded signal
we would expect this to be two strings ("(ss)"). However, nothing
ensures that the signal really provides the expected number of
parameters. It's confusing to concatenate a list of extra arguments
to a list of (an unknown number of) D-Bus arguments, as you cannot
rely on the first N arguments being D-Bus arguments. You could only rely
that the last M arguments are the extra arguments. Anyway, confusing.

Just don't do this. This flexibility is not needed nor used.