improvement(client): catch exceptions for unexpecte arguments in FirewallClient._signal_receiver()
Add a try-except around dbus_to_python(). Otherwise unexpected arguments
of a D-Bus signal will raise an exception for @handle_exceptions
decorator. When talking D-Bus, we should not treat the other peer as
trusted, even if it's the firewalld daemon.

On the other hand, move cb_args.extend(cb[1]) out of the try-except.
This really is not supposed to fail.