improvement(dbus_utils): extend dbus_to_python() and add dbus_to_python_args() helper
dbus_to_python() now also supports as expected types the D-Bus types
(for example, dbus.ObjectPath). The benefit is of course, to assert that
the type is "o" and not merely "s" (both are converted to str).

Also add dbus_to_python_args(), which extracts a duple of D-Bus
arguments, checks the types and returns the corresponding python values.