improvement(client): handle untrusted arguments from D-Bus signal in _dbus_connection_changed()
We should trust D-Bus in a certain sense, for example that when make a
D-Bus call, that we will get some sort of reply by the D-Bus library.
But we should not trust the data on D-Bus, for example that the
arguments of a signal are as we expect.

Change the signature of _dbus_connection_changed() to *dbus_args, so any
arguments can be received. And then unpack the arguments with checking
the type.

Silently ignore any unexpected arguments.

Also add an "assert name == config.dbus.DBUS_INTERFACE". Note that we
already ensure the first argument to be as expected via

        self.bus.add_signal_receiver(
            handler_function=self._dbus_connection_changed,
            signal_name="NameOwnerChanged",
            dbus_interface="org.freedesktop.DBus",
            arg0=config.dbus.DBUS_INTERFACE)