improvement(client): better handle exceptions from D-Bus callback in _dbus_receiver()
_dbus_receiver() already has a @handle_exception decorator. Adding that
decorator on every function is probably bad style, because we should
handle exceptions close to where they happen, so we can understand *why*
the happen. Having a catch-all hides bugs and encourages sloppiness.

Anyway. Since _dbus_receiver() already catches all exceptions, we can
just re-raise exceptions from cb(). Note, if cb() emits exceptions at
this point, it is a bug. cb() should not let exceptions escape but
handle them as appropriate. Re-raising is thus correct, to raise
awareness to the bug.

The only exception is if the  exception comes from an unexpected
argument list to cb(*cb_args). That would happen if the D-Bus signal
contains an unexpected number of arguments. We should treat the D-Bus
API as untrusted, and not fail upon unexpected behaviors. Instead, catch
that case and silently ignore the issue. Silently ignoring can also be a
ugly, as we may miss an actual bug. On the other hand, it's more
important to be resilient against untrusted peers and avoid spamming the
log or aborting.