test(ipset_scale,rhbz1734765,zone_priority): skip failing tests in root-less container
In a root-less container (fw-in-container), we won't have permission to
increase the socket buffer. libnftables would try to do that, but fail.

This is related to [1] and /proc/sys/net/core/{rmem_max,wmem_max,wmem_default}
sysctl.

Those tests fail. Skip them, when we detect that we run inside the
container.

[1] https://git.netfilter.org/nftables/commit/?id=375505a4a8068bf7cb623e18c3aedb831c17fd0e