improvement(config): revert "config/xmlschema: remove schema files"
While firewalld command line tools now have an option `--check-config`,
it's still useful to have a scheme for our XML files.

Revert the removal of the XSD schemas.

The patch alone does not yet suffice, the schemas needs to be adjusted
for the recent changes (will be done next).

This replaces/obsoletes pull request [1], because I think we should
restore the previous schema, instead of writing a new one.

This reverts commit 81d3d6a7d84bc34a18ddd4ab6153f0dafb2fa17e.

[1] https://github.com/firewalld/firewalld/pull/492

https://github.com/firewalld/firewalld/issues/488