improvement(rpm): import dist-git files from Fedora:rawhide for f38+
This is the specfile from dist-git Fedora/rawhide (f39), from commit [1].

There are no modifications, except the changelog is stripped. The files
will be adjusted in the following commits.

This will become a unified spec file for building firewalld RPMs like we
do on RHEL/Fedora. In fact, once this effort is complete, RHEL/Fedora
should take the specfile from here (with minimal modification). The goal
is two fold:

 - to have one place of the "best" spec file, suitable for various
   downstreams, and avoiding inconsistencies and diverging.

 - to build RPMs from upstream that are configured like downstream
   builds. In particular, to build master branch in a copr for different
   RHEL/Fedora releases.

Note that these files are taken from Fedora:rawhide, aka f39. However,
the dist-git files on f38 are very similar, with only these differences:

    @@ -24,5 +24,5 @@ BuildRequires: iptables, ebtables, ipset
     BuildRequires: python3-devel
     BuildRequires: make
    -Requires: iptables, ebtables, ipset
    +Recommends: iptables, ebtables, ipset
     Suggests: iptables-nft
     Requires(post): systemd
    @@ -82,5 +82,9 @@ Requires: %{name} = %{version}-%{release}
     Requires: firewall-config = %{version}-%{release}
     Requires: hicolor-icon-theme
    +%if (0%{?fedora} >= 39 || 0%{?rhel} >= 10)
    +Requires: python3-pyqt6
    +%else
     Requires: python3-qt5-base
    +%endif
     Requires: python3-gobject
     Requires: libnotify

This means, the spec file is suitable for firewalld master branch on f38 and
newer.

[1] https://src.fedoraproject.org/rpms/firewalld/c/4d63fa14ad0ef462713fd3b8fee80a68adca78e0